+++ CHANGELOG.md
@@ -5,13 +5,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.5.3] - 2025-06-22
+- fix #91, LATCH constants
+- update parameter constants section in .h file
+- update keywords.txt
+- update unit_tests (constants)
+- minor edits
+
 ## [0.5.2] - 2025-01-27
 - Add note about falling edge when using interrupts, see #87
 - update comments in some examples
-- minor edits 
+- minor edits
 
 ## [0.5.1] - 2024-10-17
-- fix #82, minimal timeout 10 ms for RTOS, kudos to deKees687 
+- fix #82, minimal timeout 10 ms for RTOS, kudos to deKees687
 - set error flag for TIMEOUT
 - add error codes to keywords.txt